apps/server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg[binary]==3.1.13
alembic==1.13.1
python-dotenv==1.0.0
httpx==0.25.2